Added Font Scaling Factor
Added new parameter to the "specifyUniversalFont" function, a float called newFontScalingFactor. This value is multiplied to the original font's size when creating a replacement font. The purpose of this is to allow the developer to scale fonts (that are different pixel sizes at the same 'size') so that replacement fonts match the original font's size more closely.